+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changelog
 Notable changes:
 
 * New interactive `ubik`, which will use one found on `PATH` as a last resort.
+* `bashy-core`:
+  * `arg-processor`: New post-process hook `post-process-args-call`.
 * `bashy-basics`:
   * `buildy`:
     * New command `buildy make-distro`, cribbed from sibling project `lactoserv`
@@ -15,7 +17,7 @@ Notable changes:
   * New command `ls-files`, top-level and non-git-specific (as opposed to the
     one in `buildy`).
   * Renamed `jbash-array` to `jset-array`, and made major improvements.
-  * New commadn `jstring`.
+  * New command `jstring`.
 * New sub-library `bashy-node`, for Node-specific stuff, particularly help
   building and maintaining Node projects. As with `make-distro` (above), this
   started out life in the sibling project `lactoserv`. **Nore:** This